+++ server/openapi_server/controllers/person_controller.py
@@ -2,21 +2,19 @@ import connexion
 from mongoengine.errors import DoesNotExist, NotUniqueError
 from mongoengine.queryset.visitor import Q
 
-from openapi_server.dbmodels.person import Person as DbPerson  # noqa: E501
+from openapi_server.dbmodels.person import Person as DbPerson
 from openapi_server.dbmodels.organization import Organization as DbOrganization  # noqa: E501
-from openapi_server.models.error import Error  # noqa: E501
-from openapi_server.models.page_of_persons import PageOfPersons  # noqa: E501
-from openapi_server.models.person import Person  # noqa: E501
+from openapi_server.models.error import Error
+from openapi_server.models.page_of_persons import PageOfPersons
+from openapi_server.models.person import Person
+from openapi_server.models.person_create_response import PersonCreateResponse
 from openapi_server.config import Config
 
 
-def create_person(person=None):  # noqa: E501
+def create_person():
     """Create a person
 
-    Create a person with the specified name # noqa: E501
-
-    :param person:
-    :type person: dict | bytes
+    Create a person with the specified name
 
     :rtype: Person
     """
@@ -34,20 +32,22 @@ def create_person(person=None):  # noqa: E501
                     res = Error(
                         f'The organization {org_id} was not found',
                         status)
+                    return res, status
+
             # create the person
-            if status is None:
-                try:
-                    db_person = DbPerson(
-                        firstName=person.first_name,
-                        lastName=person.last_name,
-                        email=person.email,
-                        organizations=person.organizations
-                    ).save(force_insert=True)
-                    res = Person.from_dict(db_person.to_dict())
-                    status = 200
-                except NotUniqueError as error:
-                    status = 409
-                    res = Error("Conflict", status, str(error))
+            try:
+                db_person = DbPerson(
+                    firstName=person.first_name,
+                    lastName=person.last_name,
+                    email=person.email,
+                    organizations=person.organizations
+                ).save(force_insert=True)
+                new_id = db_person.to_dict().get("personId")
+                res = PersonCreateResponse(person_id=new_id)
+                status = 201
+            except NotUniqueError as error:
+                status = 409
+                res = Error("Conflict", status, str(error))
         else:
             status = 400
             res = Error("Bad request", status)
@@ -58,10 +58,10 @@ def create_person(person=None):  # noqa: E501
     return res, status
 
 
-def delete_person(person_id):  # noqa: E501
+def delete_person(person_id):
     """Delete a person
 
-    Deletes the person specified # noqa: E501
+    Deletes the person specified
 
     :param person_id: The ID of the person
     :type person_id: str
@@ -72,23 +72,23 @@ def delete_person(person_id):  # noqa: E501
     status = None
     try:
         db_person = DbPerson.objects(personId=person_id).first()
-        res = Person.from_dict(db_person.to_dict())
-        db_person.delete()
-        status = 200
-    except DoesNotExist:
-        status = 404
-        res = Error("The specified resource was not found", status)
+        if db_person:
+            db_person.delete()
+            res = {}
+            status = 200
+        else:
+            status = 404
+            res = Error("The specified resource was not found", status)
     except Exception as error:
         status = 500
         res = Error("Internal error", status, str(error))
-
     return res, status
 
 
-def get_person(person_id):  # noqa: E501
+def get_person(person_id):
     """Get a person
 
-    Returns the person specified # noqa: E501
+    Returns the person specified
 
     :param person_id: The ID of the person
     :type person_id: str
@@ -99,11 +99,12 @@ def get_person(person_id):  # noqa: E501
     status = None
     try:
         db_person = DbPerson.objects(personId=person_id).first()
-        res = Person.from_dict(db_person.to_dict())
-        status = 200
-    except DoesNotExist:
-        status = 404
-        res = Error("The specified resource was not found", status)
+        if db_person:
+            res = Person.from_dict(db_person.to_dict())
+            status = 200
+        else:
+            status = 404
+            res = Error("The specified resource was not found", status)
     except Exception as error:
         status = 500
         res = Error("Internal error", status, str(error))
@@ -111,10 +112,10 @@ def get_person(person_id):  # noqa: E501
     return res, status
 
 
-def list_persons(limit=None, offset=None, filter_=None):  # noqa: E501
+def list_persons(limit=None, offset=None, filter_=None):
     """Get all persons
 
-    Returns the persons # noqa: E501
+    Returns the persons
 
     :param limit: Maximum number of results returned
     :type limit: int
@@ -135,7 +136,7 @@ def list_persons(limit=None, offset=None, filter_=None):  # noqa: E501
         organization_q = Q(organizations__contains=filter_['organization']) \
             if 'organization' in filter_ else Q()
         db_persons = DbPerson.objects(
-           first_name_q & last_name_q & email_q & organization_q
+            first_name_q & last_name_q & email_q & organization_q
         ).skip(offset).limit(limit)
         persons = [Person.from_dict(d.to_dict()) for d in db_persons]
         next_ = ""
@@ -148,11 +149,12 @@ def list_persons(limit=None, offset=None, filter_=None):  # noqa: E501
             links={
                 "next": next_
             },
+            total_results=len(persons),
             persons=persons)
         status = 200
-    except DoesNotExist:
-        status = 404
-        res = Error("The specified resource was not found", status)
+    except DoesNotExist:  # TODO: update exception handling
+        status = 400
+        res = Error("Bad request", status)
     except Exception as error:
         status = 500
         res = Error("Internal error", status, str(error))
